+++ pkg/artifactory/resource/repository/repository_test.go
@@ -455,3 +455,63 @@ func TestAccRepository_invalid_key(t *testing.T) {
 		},
 	})
 }
+
+func TestAccRepository_unknown_project_environments(t *testing.T) {
+    projectKey := fmt.Sprintf("t%d", testutil.RandomInt())
+    repoName := fmt.Sprintf("%s-generic-local", projectKey)
+    environment := fmt.Sprintf("prjenv%d", testutil.RandomInt())
+
+    _, fqrn, name := testutil.MkNames(repoName, "artifactory_local_generic_repository")
+
+    params := map[string]interface{}{
+        "name":        name,
+        "projectKey":  projectKey,
+        "environment": environment,
+    }
+    localRepositoryBasic := util.ExecuteTemplate("TestAccLocalGenericRepository", `
+        resource "project" "{{ .projectKey }}" {
+            key = "{{ .projectKey }}"
+            display_name = "{{ .projectKey }}"
+            description  = "My Project"
+            admin_privileges {
+                manage_members   = true
+                manage_resources = true
+                index_resources  = true
+            }
+            max_storage_in_gibibytes   = 10
+            block_deployments_on_limit = false
+            email_notification         = true
+        }
+
+        resource "project_environment" "{{ .environment }}" {
+          name        = "{{ .environment }}"
+          project_key = project.{{ .projectKey }}.id
+        }
+
+        resource "artifactory_local_generic_repository" "{{ .name }}" {
+          key                  = "{{ .name }}"
+          project_key          = project.{{ .projectKey }}.key
+          project_environments = [ project_environment.{{ .environment }}.id ]
+        }
+    `, params)
+
+    resource.Test(t, resource.TestCase{
+        PreCheck: func() { acctest.PreCheck(t) },
+        ExternalProviders: map[string]resource.ExternalProvider{
+            "project": {
+                Source: "jfrog/project",
+            },
+        },
+        ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+        CheckDestroy:             acctest.VerifyDeleted(t, fqrn, "key", acctest.CheckRepo),
+        Steps: []resource.TestStep{
+            {
+                Config:      localRepositoryBasic,
+                Check: resource.ComposeTestCheckFunc(
+                    resource.TestCheckResourceAttr(fqrn, "project_environments.#", "1"),
+                    resource.TestCheckResourceAttr(fqrn, "project_environments.0", fmt.Sprintf("%s-%s",projectKey,environment)),
+                ),
+            },
+        },
+    })
+}