+++ CHANGELOG.md
@@ -4,6 +4,7 @@ BUG FIXES:
 
 * resource/artifactory_\*\_custom_webhook: Fix secret name validation to validate keys instead of values in the secrets map. The validation now correctly checks that secret names follow the pattern `^[a-zA-Z_][a-zA-Z0-9_]*$` instead of validating the secret values. Issue: [#1297](https://github.com/jfrog/terraform-provider-artifactory/issues/1297). PR: [#1298](https://github.com/jfrog/terraform-provider-artifactory/pull/1298)
 * resource/artifactory_package_cleanup_policy: Fix re-add project_key for artifactory_package_cleanup_policy. Issue: [#1301](https://github.com/jfrog/terraform-provider-artifactory/issues/1301). PR: [#1302](https://github.com/jfrog/terraform-provider-artifactory/pull/1302)
+* resource/repository/repository.go: Use types.String to support "unknown" project_environments elements. Issue: [#1251](https://github.com/jfrog/terraform-provider-artifactory/issues/1251). PR: [#1252](https://github.com/jfrog/terraform-provider-artifactory/pull/1252)
 
 IMPROVEMENTS: